+++ packages/Zooniverse-React-Components/lib/index.js
@@ -148,6 +148,15 @@ Object.defineProperty(exports, 'LogoutButton', {
   }
 });
 
+var _mobileNavigationMenu = require('./components/layout/mobile-navigation-menu');
+
+Object.defineProperty(exports, 'MobileNavigationMenu', {
+  enumerable: true,
+  get: function get() {
+    return _interopRequireDefault(_mobileNavigationMenu).default;
+  }
+});
+
 var _oauthModal = require('./components/layout/oauth-modal');
 
 Object.defineProperty(exports, 'OauthModal', {
@@ -166,6 +175,24 @@ Object.defineProperty(exports, 'OauthGoogleIcon', {
   }
 });
 
+var _signedInUserNavigation = require('./components/layout/signed-in-user-navigation');
+
+Object.defineProperty(exports, 'SignedInUserNavigation', {
+  enumerable: true,
+  get: function get() {
+    return _interopRequireDefault(_signedInUserNavigation).default;
+  }
+});
+
+var _signedOutUserNavigation = require('./components/layout/signed-out-user-navigation');
+
+Object.defineProperty(exports, 'SignedOutUserNavigation', {
+  enumerable: true,
+  get: function get() {
+    return _interopRequireDefault(_signedOutUserNavigation).default;
+  }
+});
+
 var _userMenu = require('./components/layout/user-menu');
 
 Object.defineProperty(exports, 'UserMenu', {
@@ -181,6 +208,15 @@ Object.defineProperty(exports, 'UserNavigation', {
   enumerable: true,
   get: function get() {
     return _interopRequireDefault(_userNavigation).default;
+  }
+});
+
+var _withMobileView = require('./components/layout/with-mobile-view');
+
+Object.defineProperty(exports, 'WithMobileView', {
+  enumerable: true,
+  get: function get() {
+    return _interopRequireDefault(_withMobileView).default;
   }
 });