+++ packages/Zooniverse-React-Components/src/components/layout/login-button.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Button from 'grommet/components/Button';
 
-const LoginButton = ({ className, label, login, plain, toggleModal }) => {
+export default function LoginButton({ className, label, login, plain, toggleModal }) {
   return (
     <Button type="button" className={className} onClick={login || toggleModal} label={label} plain={plain} />
   );
@@ -23,5 +23,3 @@ LoginButton.propTypes = {
   plain: PropTypes.bool,
   toggleModal: PropTypes.func
 };
-
-export default LoginButton;