+++ packages/Zooniverse-React-Components/src/components/layout/user-navigation.jsx
@@ -3,32 +3,50 @@ import PropTypes from 'prop-types';
 import Anchor from 'grommet/components/Anchor';
 import Menu from 'grommet/components/Menu';
 
-const UserNavigation = (props) => {
-  const createKeyedAnchorItem = (navItem, i) => {
-    return (React.cloneElement(navItem, { key: `navItem-${i}` }));
-  };
+import withMobileView from './with-mobile-view';
+
+export function UserNavigation(props) {
+  const messagesLabel = props.isMobile ?
+    <i className="fa fa-bell-o fa-fw" aria-hidden="true" aria-label={props.messagesLabel} /> :
+    props.messagesLabel;
+  const notificationsLabel = props.isMobile ?
+   <i className="fa fa-envelope-o fa-fw" aria-hidden="true" aria-label={props.notificationsLabel} /> :
+   props.notificationsLabel
 
   return (
     <Menu className="user-navigation" align="center" direction="row" size="small" responsive={false} inline={true}>
-      {props.userNavigationNavList.map((navItem, i) => {
-        return createKeyedAnchorItem(navItem, i);
-      })}
+      {React.cloneElement(props.notificationsLink, { label: notificationsLabel })}
+      {React.cloneElement(props.messagesLink, { label: messagesLabel })}
     </Menu>
   );
 }
 
 UserNavigation.defaultProps = {
-  userNavigationNavList: [
-    <Anchor className="zoo-header__link--small" href="http://www.zooniverse.org/notifications" label="Notifications" />,
-    <Anchor className="zoo-header__link--small" href="http://www.zooniverse.org/inbox" label="Messages" />
-  ]
+  isMobile: false,
+  messagesLabel: "Messages",
+  notificationsLabel: "Notifications",
+  messagesLink: <Anchor className="zoo-header__link--small" href="http://www.zooniverse.org/inbox" />,
+  notificationsLink: <Anchor className="zoo-header__link--small" href="http://www.zooniverse.org/notifications" />
 };
 
 UserNavigation.propTypes = {
-  userNavigationNavList: PropTypes.arrayOf(PropTypes.oneOfType([
+  isMobile: PropTypes.bool,
+  messagesLabel: PropTypes.oneOfType([
     PropTypes.node,
     PropTypes.string
-  ])).isRequired
+  ]).isRequired,
+  messagesLink: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.string
+  ]).isRequired,
+  notificationsLabel: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.string
+  ]).isRequired,
+  notificationsLink: PropTypes.oneOfType([
+    PropTypes.node,
+    PropTypes.string
+  ]).isRequired
 };
 
-export default UserNavigation;
+export default withMobileView(UserNavigation);